Redefine `ModelOp` in terms of a state monad transformer.
The `ModelOp` type was originally defined as:

    ```
    PoolDatabase -> (Either PoolErr a, PoolDatabase)
    ```

Note that this is very similar to:

    ```
    PoolDatabase -> Either PoolErr (a, PoolDatabase)
    ```

The latter definition has the advantage that if an operation fails, it fails
completely: the caller cannot see a partially-updated database state. In the
vast majority of cases, this is what we actually want: it's extremely rare
that we want to retain the results of a partial update.

Note that the latter definition has the same shape as the type of `runStateT`,
from `StateT`:

    ```
    runStateT :: s -> m (a, s)
    ```

Let's for a moment redefine `ModelOp` in terms of `StateT`:

    ```
    StateT PoolDatabase (Either PoolErr) a
    ```

The type of `runStateT` then becomes:

    ```
    runStateT :: PoolDatabase -> Either PoolErr (a, PoolDatabase)
    ```

Which is exactly the same as our latter definition above.

By redefining `ModelOp` in terms of `StateT`, it's now possible to express
model operations monadically. This allows for a more concise style, with
less boilerplate.

A simple example:

    ```
    -mCleanDatabase _ = (Right (), emptyPoolDatabase)
    -mCleanDatabase = State.put emptyPoolDatabase
    ```

A more complex example:

    ```
    mRollbackTo ti point = do
        modify #distributions
            $ Map.mapMaybeWithKey $ discardBy $ runIdentity . ti . epochOf
        modify #pools
            $ Map.filter (not . L.null) . fmap (filter ((<= point) . slotNo))
        modify #registrations
            $ Map.mapMaybeWithKey $ discardBy id . view #slotNo . fst
        modify #retirements
            $ Map.mapMaybeWithKey $ discardBy id . view #slotNo . fst
        poolIds <-
            Set.fromList <$> mListRegisteredPools
        modify #owners
            $ flip Map.restrictKeys poolIds
    ```